Move dataclass option tests into test_nested_options.py
